fix: silence -Wformat-truncation in pronto_string
Cast pairs.size()/2 to uint16_t so GCC knows it fits in %04X.

+++ components/samsung_aqv/protocol.h
@@ -161,7 +161,7 @@ inline void append_bits_pronto(std::vector<uint16_t> &pairs, const uint8_t *bits
 inline std::string pronto_string(uint16_t freq, const std::vector<uint16_t> &pairs) {
   std::string result;
   char buf[6];
-  int n = pairs.size() / 2;
+  uint16_t n = static_cast<uint16_t>(pairs.size() / 2);
   snprintf(buf, sizeof(buf), "%04X", 0);
   result += buf;
   result += ' ';
